+++ src/common/types/property-bag/text-alternative-property-bag.ts
